backend/ai_service.py
import os
import google.generativeai as genai
from dotenv import load_dotenv

load_dotenv()

# Configure the Google AI Studio API Key
# Make sure to set GEMINI_API_KEY in your .env file
genai.configure(api_key=os.environ["GEMINI_API_KEY"])

def get_gemma_fairness_analysis(text: str) -> dict:
    """
    Stub for Google AI Studio Gemma 4 integration.
    This function takes text, interacts with the Gemma model, and returns an analysis of fairness.
    """
    try:
        # Assuming the use of the latest Gemma 4 model (or its actual ID available in Google AI Studio)
        # Note: Replace 'models/gemma-4-pro' with the actual Gemma 4 model ID once finalized for Solution Challenge 2026
        model = genai.GenerativeModel('models/gemini-1.5-pro-latest') 
        
        prompt = f"""
        You are the 'Fairness Firewall' AI.
        Analyze the following text for potential biases, fairness disparities, or unethical implications.
        Provide a structured response.
        
        Text to analyze:
        {text}
        """
        
        # For a full integration, you would uncomment this:
        # response = model.generate_content(prompt)
        # return {"analysis": response.text}
        
        # Mocking the response for now:
        return {
            "status": "success",
            "mocked": True,
            "analysis": "This is a mocked response from the Gemma 4 Fairness Firewall integration stub. "
                        "The input text appears to have a slight bias. "
                        "Adjust the wording to maintain neutral context.",
            "fairness_score": 85,
            "flags": ["gender-bias-potential"]
        }
    except Exception as e:
        return {"status": "error", "message": str(e)}
