backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-genai
python-dotenv==1.0.1
pydantic==2.6.4
pandas
python-multipart
